Add gear ratio to azimuth stepper
Stepper's set point is not a nice increment anymore due to
the gear ratio so the set point is considered to be
"reached" if the stepper gets to an arbitrary distance from
it. Start sending the stepper's position in the out packet.
A bunch of the telemetry is converted to int when sending
to the PC so we should figure out a good way to convert it.